UIController now derives from Selectable
This ends up simplifying a lot of code, but this also means I have to use the left mouse button to drag instead of the middle mouse button. This is because the base class, Selectable, implements the graphics for selecting for only the left mouse button, and I don't feel like changing it to suit my wants. The change of controls is also reflected in the input actions for the Cubism model.