Create window manager and various other changes
The window manager is now responsible for handling the clickthrough of a window via the IWindowClickthroughHandler interface. Before, the handling was done via a unity event. This, however, had a bug in which the clickthrough state of the window could be overridden by dragging game objects over and out of each other, causing the window to become clickthrough despite the pointer still hovering over the most recently dragged game object. With the new implementation, all clickthrough handlers are taken into account, instead of one at a time. This is done by storing all game objects that implement the aforementioned interface in an IEnumerable, and checking to see if at least one element sets the window's clickthrough to false.

Various other changes:
- SetClickthrough now gets and does bitwise operations on the current window style, as opposed to a hardcoded style
- Generalized transparent window game object
- Made custom editor menu items for window game objects
- Removed unnecessary #if directives
- Organized folder structure and namespaces
- Deleted TypewriterEffect for now
- Tweaked input actions a bit
- Replaced soon-to-be deprecated functions