v1.0.0-alpha.3 (unfinished & unreleased)
I'm not actually bothered and/or too lazy to figure out how to get the python end of Wanko to properly communicate with the C# end the way I initially wanted to, anymore. This was just a proof of concept to show that I could have them at least communicate in some minimal way, and have the python embed, pip packages, and other dependencies automatically install and handle themselves. It'd be easier and more practical to just use the Unity package Hugging Face offers, along with the .NET binding for Character AI, but it was fun to experiment with python integration in Unity anyway.

Changes from v1.0.0-alpha.2:
- Full python integration (may come back to this commit if I ever need to actually use python in the future)
- PyWanko (the python end of Wanko, which actually works as a standalone, but I could figure out how to seamlessly integrate it using Python.NET)
- Lots of refactoring, reorganizing, touching up, and other small tweaks to everything here and there
- .gitignore'd the architectural-dependent python embed, which also bloated the repository (it now installs itself upon project import)